refactor: groupByDate, calcTotal 함수 util로 뺌
