+++ client/src/utils/index.ts
@@ -1,0 +1,3 @@
+export { calcTotal } from './calc-total';
+export { groupByDate } from './group-by-date';
+export { insertAt } from './insert-item-at';
